# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Patrick's website API for Certification listings",
        "title": "Certification Listings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "container-app-ilm34snayo7qw.politeplant-0cb21c20.australiaeast.azurecontainerapps.io",
    "basePath": "/certifications/api/v1",
    "paths": {
        "/cert/{key}/{value}": {
            "get": {
                "description": "Get a certification by a certain key \u0026 value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certifications"
                ],
                "summary": "Show a Specified Cert",
                "parameters": [
                    {
                        "enum": [
                            "id",
                            "category",
                            "company"
                        ],
                        "type": "string",
                        "description": "Select Category for Search",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Microsoft",
                        "description": "Select Category for Search",
                        "name": "value",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Certs"
                        }
                    }
                }
            }
        },
        "/certs": {
            "get": {
                "description": "List all current Certifications by Patrick",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certifications"
                ],
                "summary": "Show Certification Listings",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 3,
                        "description": "Top results",
                        "name": "top",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Certs"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Certs": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "skills": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}