# src/item.go
package main

// <model>
type Certs struct {
	Id       string `json:"id"`
	Category string `json:"category"`
	Compnay  string `json:"company"`
	Title    string `json:"title"`
	Date     string `json:"date"`
	Link     string `json:"link"`
	Skills    string `json:"skills"`
}